requirements.txt
# requirements.txt (öneri)
Django==4.2.7
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.5
asgiref==3.8.1
requests==2.32.3
